move CatsEffect into testing module
